requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
boilerpy3==1.0.7
boto3==1.34.111
botocore==1.34.111
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
docx==0.2.4
elastic-transport==8.13.0
elasticsearch==8.13.1
elasticsearch-haystack==0.4.0
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
h11==0.14.0
haystack-ai==2.1.2
haystack-bm25==1.0.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
lazy-imports==0.3.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.1
openpyxl==3.1.2
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pillow==10.4.0
posthog==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMuPDF==1.24.5
pypdf==4.2.0
PySocks==1.7.1
pytesseract==0.3.8
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
playwright
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
selenium==4.21.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
tenacity==8.3.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uuid==1.30
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
nest_asyncio
oauth2client
tldextract==5.1.2

markdownify==0.12.1
schedule==1.2.2

scikit-learn==1.5.1

nltk==3.8.1
pyjwt==2.4.0

PaddleOCR==2.6.1.3
Pillow==10.0.0
pandas==2.0.3
tabulate==0.9.0